+++ lightcurve/src/core/repository/models/probability.py
@@ -1,0 +1,11 @@
+from typing import Optional
+
+from pydantic import BaseModel
+
+
+class Probability(BaseModel):
+    classifier_name: str
+    classifier_version: str
+    class_name: str
+    probability: float
+    ranking: int
